+++ CHANGELOG.md
@@ -26,7 +26,7 @@ The following changes have been implemented but not released yet:
   this bug in other functions based on this one, such as `getPodUrlAll`.
 
 - `getPodUrlAll` no longer throws if the WebID only appears as the object in an
-alternative profile (and not as a subject), which is a valid case.
+  alternative profile (and not as a subject), which is a valid case.
 
 ## [1.19.0] - 2022-02-09